requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-auth==2.39.0
google-genai==1.11.0
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
pillow>=9.0.0
playwright>=1.30.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
asyncio>=3.4.3
python-dotenv>=1.0.0
pandas>=2.0.0
